Just clamp and trunc left
